hickory_resolver

Module config

Configuration for a resolver

Structs

NameServerConfig

Configuration for the NameServer

NameServerConfigGroup

A set of name_servers to associate with a ResolverConfig.

ResolverConfig

Configuration for the upstream nameservers to use for resolution

ResolverOpts

Configuration for the Resolver

Enums

LookupIpStrategy

The lookup ip strategy

ResolveHosts

Whether the system hosts file should be respected by the resolver.

ServerOrderingStrategy

The strategy for establishing the query order of name servers in a pool.

Constants

CLOUDFLARE_IPS

IP addresses for Cloudflare’s 1.1.1.1 DNS service

GOOGLE_IPS

IP addresses for Google Public DNS

QUAD9_IPS

IP address for the Quad9 DNS service