hickory_resolver

Module lookup

Lookup result from a resolution of ipv4 and ipv6 records with a Resolver.

Structs

Lookup

Result of a DNS query when querying for any record type supported by the Hickory DNS Proto library.

LookupIter

Borrowed view of set of RDatas returned from a Lookup

LookupRecordIter

Borrowed view of set of Records returned from a Lookup

SrvLookup

The result of an SRV lookup

SrvLookupIter

An iterator over the Lookup type

TypedLookup

Contains the results of a lookup for the associated RecordType

TypedLookupIter

An iterator over the Lookup type