hickory_resolver
Module name_server

A module with associated items for working with nameservers

Structs

GenericConnection

A connected DNS handle

GenericConnector

Default connector for GenericConnection

NameServer

This struct is used to create DnsHandle with the help of P.

NameServerPool

Abstract interface for mocking purpose

TokioHandle

A handle to the Tokio runtime

TokioRuntimeProvider

The Tokio Runtime for async execution

Traits

ConnectionProvider

Create DnsHandle with the help of RuntimeProvider. This trait is designed for customization.

RuntimeProvider

RuntimeProvider defines which async runtime that handles IO and timers.

Spawn

A type defines the Handle which can spawn future.

Type Aliases

GenericNameServer

Specifies the details of a remote NameServer used for lookups

GenericNameServerPool

A pool of NameServers

TokioConnectionProvider

Default ConnectionProvider with GenericConnection.