Struct hickory_resolver::name_server::NameServer

pub struct NameServer<P: ConnectionProvider> { /* private fields */ }

This struct is used to create DnsHandle with the help of P.

Implementations

impl<P> NameServer<P>
where P: ConnectionProvider + Send,

pub fn new( config: NameServerConfig, options: ResolverOpts, connection_provider: P ) -> Self

Construct a new Nameserver with the configuration and options. The connection provider will create UDP and TCP sockets

pub fn trust_nx_responses(&self) -> bool

Specifies that this NameServer will treat negative responses as permanent failures and will not retry