hickory_resolver

Struct ResolverBuilder

pub struct ResolverBuilder<P> { /* private fields */ }

A builder to construct a Resolver.

Created by Resolver::builder.

Implementations

impl<P> ResolverBuilder<P>
where P: ConnectionProvider,

pub fn with_options(self, options: ResolverOpts) -> Self

Sets the ResolverOpts to be used by the resolver.

NB: A ResolverBuilder<P> will use the system configuration e.g., resolv.conf, by default. Usage of this method will overwrite any options set by the system configuration.

See system_conf for functions that can parse a ResolverOpts from the system configuration, or use options_mut() to acquire a muitable reference to the existing ResolverOpts.

pub fn options_mut(&mut self) -> &mut ResolverOpts

Returns a mutable reference to the ResolverOpts.

pub fn build(self) -> Resolver<P>

Construct the resolver.