Module hickory_resolver::lookup

Lookup result from a resolution of ipv4 and ipv6 records with a Resolver.

Structs

Ipv4Lookup

Contains the results of a lookup for the associated RecordType

Ipv4LookupIntoIter

Borrowed view of set of RDatas returned from a Lookup

Ipv4LookupIter

An iterator over the Lookup type

Ipv6Lookup

Contains the results of a lookup for the associated RecordType

Ipv6LookupIntoIter

Borrowed view of set of RDatas returned from a Lookup

Ipv6LookupIter

An iterator over the Lookup type

Lookup

Result of a DNS query when querying for any record type supported by the Hickory DNS Proto library.

LookupIntoIter

Borrowed view of set of RDatas returned from a Lookup.

LookupIter

Borrowed view of set of RDatas returned from a Lookup

LookupRecordIter

Borrowed view of set of Records returned from a Lookup

MxLookup

Contains the results of a lookup for the associated RecordType

MxLookupIntoIter

Borrowed view of set of RDatas returned from a Lookup

MxLookupIter

An iterator over the Lookup type

NsLookup

Contains the results of a lookup for the associated RecordType

NsLookupIntoIter

Borrowed view of set of RDatas returned from a Lookup

NsLookupIter

An iterator over the Lookup type

ReverseLookup

Contains the results of a lookup for the associated RecordType

ReverseLookupIntoIter

Borrowed view of set of RDatas returned from a Lookup

ReverseLookupIter

An iterator over the Lookup type

SoaLookup

Contains the results of a lookup for the associated RecordType

SoaLookupIntoIter

Borrowed view of set of RDatas returned from a Lookup

SoaLookupIter

An iterator over the Lookup type

SrvLookup

The result of an SRV lookup

SrvLookupIntoIter

Borrowed view of set of RDatas returned from a Lookup

SrvLookupIter

An iterator over the Lookup type

TlsaLookup

Contains the results of a lookup for the associated RecordType

TlsaLookupIntoIter

Borrowed view of set of RDatas returned from a Lookup

TlsaLookupIter

An iterator over the Lookup type

TxtLookup

Contains the results of a lookup for the associated RecordType

TxtLookupIntoIter

Borrowed view of set of RDatas returned from a Lookup

TxtLookupIter

An iterator over the Lookup type