Type Alias hickory_resolver::ResolverFuture

pub type ResolverFuture = TokioAsyncResolver;

👎Deprecated: use [hickory_resolver::AsyncResolver] instead

This is an alias for AsyncResolver, which replaced the type previously called ResolverFuture.

Note

For users of ResolverFuture, the return type for ResolverFuture::new has changed since version 0.9 of hickory-resolver. It now returns a tuple of an AsyncResolver and a background future, which must be spawned on a reactor before any lookup futures will run.

See the AsyncResolver documentation for more information on how to use the background future.

Aliased Type

struct ResolverFuture { /* private fields */ }